refactor: decouple shared schema dependencies
