+++ src/shared/schemas/index.ts
@@ -1,9 +1,16 @@
+// Core schemas (no external dependencies)
+export * from './agent';
+export * from './diff';
 export * from './errors';
 export * from './identifiers';
-export * from './log';
-export * from './output';
-export * from './prompts';
-export * from './stream';
-export * from './taskGroup';
 export * from './todo';
 export * from './usage';
+
+// Log & stream schemas
+export * from './log';
+export * from './stream';
+export * from './taskGroup';
+
+// Output & prompt schemas
+export * from './output';
+export * from './prompts';
